Apps2pay/MINT/mitra-mint/requirements.txt
# Core testing libraries
Appium-Python-Client
pytest
pytest-html

# Env variable management
python-dotenv

# Report testing
allure-pytest

# Pararel testing
pytest-xdist